Packed code into smaller modules. Renamed bot.py to main.py
